requirements.txt
fastapi
alembic
uvicorn[standard]
sqlalchemy
python-docx
openai
anthropic
google-generativeai
python-multipart
aiofiles
bcrypt
